refactor(utils): reorganize note types and imports in writer module
