+++ packages/utils/src/types.d.ts
@@ -1,0 +1,5 @@
+declare module '*.md' {
+  const content: string
+
+  export default content
+}
