+++ packages/utils/src/writer/index.ts
@@ -1,3 +1,3 @@
+export * from '../types/notes.types'
 export { enhanceBulletPoint, enhanceBulletPoints } from './enhance-bullet-point'
 export { parseNoteDetails } from './notes'
-export type { BulletPoint, NoteDetails } from './text.d.ts'
